% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_abundance_trends.R
\name{plot_abundance_trends}
\alias{plot_abundance_trends}
\title{Plot Yearly Trends in Abundance}
\usage{
plot_abundance_trends(
  hca_data,
  abundance_data,
  focal_species,
  add_stats = TRUE
)
}
\arguments{
\item{hca_data}{A data frame produced by `calculate_hca()`.}

\item{abundance_data}{A standardized data frame from `get_plant_data(output = "standardized")`.}

\item{focal_species}{A character string specifying the plant code of the focal species.}

\item{add_stats}{A logical value. If `TRUE` (the default), statistical comparisons
are performed and p-values are added to the plot.}
}
\value{
A ggplot object.
}
\description{
Creates faceted boxplots to visualize year-to-year changes in conspecific
and heterospecific (HCA) abundance, with an option for statistical comparisons.
}
\details{
This function visualizes the distribution of abundance values for each year.
If `add_stats = TRUE`, it uses `ggpubr::stat_compare_means` to perform a
Wilcoxon rank-sum test (Mann-Whitney U test) to check for significant
differences between consecutive years.
}
