% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plant_sharing_network.R
\name{plot_plant_sharing_network}
\alias{plot_plant_sharing_network}
\title{Plot a Plant-Centric Pollinator Sharing Network}
\usage{
plot_plant_sharing_network(
  size_by = "interactions",
  focal_plant = NULL,
  plant_abundance_data = NULL,
  colors = list(plant = "lightseagreen", edge = rgb(104, 99, 219, max = 255, alpha = 80),
    focal_plant = "gold", focal_edge = "red"),
  ...
)
}
\arguments{
\item{size_by}{A character string: `"interactions"` (default) or `"abundance"`.}

\item{focal_plant}{A character string specifying a plant code to highlight.}

\item{plant_abundance_data}{A data frame of standardized plant abundance.}

\item{colors}{A named list of colors for the plot components.}

\item{...}{Filtering arguments passed to `get_interaction_data()`.}
}
\value{
A plot generated by `igraph`.
}
\description{
This function visualizes the pollinator sharing network between plant species
in a circular layout.
}
