% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollen_network_data}
\alias{pollen_network_data}
\title{Pollen Co-deposition Network Data}
\format{
A data frame with one row per pollen type per stigma.
\describe{
  \item{stigma_id}{A unique identifier for each stigma.}
  \item{year}{The year of collection.}
  \item{inflorescence_id}{Identifier for the inflorescence.}
  \item{plant_code}{The code for the plant species from which the pollen originated.}
  \item{pollen_count}{The number of pollen grains of that type on the stigma.}
}
}
\usage{
pollen_network_data
}
\description{
A tidy data frame of pollen counts from Succisa pratensis stigmas, showing
which plant pollen types co-occur.
}
\keyword{datasets}
